move gitimport resource to core
