+++ services/orchest-webserver/app/app/views/background_tasks.py
@@ -1,8 +1,6 @@
 from flask_restful import Api, Resource
 from flask import request
 import os
-import subprocess
-from subprocess import Popen
 import uuid
 
 from app.models import BackgroundTask
@@ -39,38 +37,6 @@ def register_background_tasks_view(app, db):
 
             return background_task_schema.dump(task)
 
-    class ImportGitProjectListResource(Resource):
-        def post(self):
-            n_uuid = str(uuid.uuid4())
-            new_task = BackgroundTask(
-                task_uuid=n_uuid, task_type="GIT_CLONE_PROJECT", status="PENDING"
-            )
-            db.session.add(new_task)
-            db.session.commit()
-
-            # start the background process in charge of cloning
-            file_dir = os.path.dirname(os.path.realpath(__file__))
-            background_task_process = Popen(
-                [
-                    "python3",
-                    "-m",
-                    "scripts.background_tasks",
-                    "--type",
-                    "git_clone_project",
-                    "--uuid",
-                    n_uuid,
-                    "--url",
-                    request.json["url"],
-                    "--path",
-                    request.json["project_name"],
-                ],
-                cwd=os.path.join(file_dir, "../.."),
-                stderr=subprocess.STDOUT,
-            )
-
-            return background_task_schema.dump(new_task)
-
-    api.add_resource(ImportGitProjectListResource, "/async/import-git")
     api.add_resource(
-        BackgroundTaskResource, "/async/background-task/<string:task_uuid>"
+        BackgroundTaskResource, "/async/background-tasks/<string:task_uuid>"
     )
